refactor: minimize code sprawl by maintaining app foundation in 1 file
